requirements.txt
requests==2.31.0
jmespath==1.0.1
pyyaml==6.0.1
python-dotenv==1.0.0 